Make quickFetch a module for include in app.js.
